Move fungi to separate module
